% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binseg.R
\name{binseg}
\alias{binseg}
\title{Change Point Detection using binseg algorithm}
\usage{
binseg(
  data,
  type_changement = "mean",
  Q = 100,
  penalty = "Manual",
  pen.value = var(data) * log(length(data))
)
}
\arguments{
\item{data}{Numeric vector containing the data to segment.}

\item{type_changement}{Le type de changement à détecter. Peut être "mean" pour le changement de moyenne, "var" pour le changement de variance, ou "meanvar" pour les deux.}

\item{Q}{Le paramètre Q pour le nombre maximal de rupture.}

\item{penalty}{Le paramètre de pénalité pour l'algorithme PELT.}

\item{pen.value}{La valeur de la pénalité.}
}
\value{
Un vecteur contenant les points de changement identifiés.
}
\description{
Change Point Detection using binseg algorithm
}
\examples{
data <- c(rnorm(40,0,2),rnorm(60,5,2),rnorm(60,1,2),rnorm(40,5,2))
type_changement = "mean"
Q = 100
penalty = "Manual"
pen.value = var(data) * log(length(data))
binseg(data, type_changement,Q, penalty, pen.value)
}
