% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPM.R
\name{cpm}
\alias{cpm}
\title{Change Point Detection using cpm algorithm}
\usage{
cpm(
  data,
  cpmType = "Student",
  ARL0 = ceiling(0.5 * length(data)/1000) * 1000,
  startup = 20
)
}
\arguments{
\item{data}{Numeric vector containing the data to segment.}

\item{cpmType}{Type of CPM algorithm to use.}

\item{ARL0}{ARL0 parameter for the CPM algorithm.}

\item{startup}{Startup parameter for the CPM algorithm.}
}
\value{
A numeric vector containing the segmentation points.
}
\description{
Change Point Detection using cpm algorithm
}
\examples{
data <- c(rnorm(40,0,2),rnorm(60,5,2),rnorm(60,1,2),rnorm(40,5,2))
cpmType = "Student"
ARL0 = ceiling(0.5 *length(data)/1000)*1000
startup = 20
cpm(data,cpmType,ARL0, startup )
}
