% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e.div.R
\name{e.div}
\alias{e.div}
\title{Change Point Detection using the e.div Algorithm}
\usage{
e.div(data, sig.lvl = 0.05, R = 199, k = NULL, min.size = 4, alpha = 1)
}
\arguments{
\item{data}{Numeric vector containing the data to segment.}

\item{sig.lvl}{The significance level for detecting change points.}

\item{R}{The maximum number of change points to detect.}

\item{k}{The number of clusters for the data. If not specified, it will be automatically determined.}

\item{min.size}{The minimum size of a segment to be considered as a change point.}

\item{alpha}{The significance level for the test of normality.}
}
\value{
A numeric vector containing the detected change points.
}
\description{
Change Point Detection using the e.div Algorithm
}
\examples{
data <- c(rnorm(40,0,2),rnorm(60,5,2),rnorm(60,1,2),rnorm(40,5,2))
sig.lvl <- 0.05
R <- 199
K <- NULL
min.size <- 4
alpha <- 1
e.div(data, sig.lvl, R, K, min.size, alpha)
}
