% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbeast.R
\name{rbeast}
\alias{rbeast}
\title{Change Point Detection using rbeast algorithm}
\usage{
rbeast(data)
}
\arguments{
\item{data}{Numeric vector containing the data to segment.}
}
\value{
Un vecteur contenant les points de changement identifiés.
}
\description{
Change Point Detection using rbeast algorithm
}
\examples{
data <- c(rnorm(40,0,2),rnorm(60,5,2),rnorm(60,1,2),rnorm(40,5,2))
rbeast(data)
}
