% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segneigh.R
\name{segneigh}
\alias{segneigh}
\title{Change Point Detection using segneigh algorithm}
\usage{
segneigh(
  data,
  type_changement = "mean",
  Q = 100,
  penalty = "Manual",
  pen.value = var(data) * log(length(data))
)
}
\arguments{
\item{data}{Numeric vector containing the data to segment.}

\item{type_changement}{Type of change to detect. Maybe "mean","var" or "meanvar".}

\item{Q}{Numeric value for the maximum number of segments.}

\item{penalty}{Method of penalty. Maybe "Manual" or "BIC" for example.}

\item{pen.value}{if penalty is "Manual", this is the value of the penalty.}
}
\value{
A numeric vector containing the segmentation points.
}
\description{
Change Point Detection using segneigh algorithm
}
\examples{
data <- c(rnorm(40,0,2),rnorm(60,5,2),rnorm(60,1,2),rnorm(40,5,2))
type_changement = "mean"
Q = 100
penalty = "Manual"
pen.value = var(data) * log(length(data))
segneigh(data, type_changement,Q, penalty, pen.value)
}
