% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PELT.R
\name{pelt}
\alias{pelt}
\title{Change Point Detection using pelt algorithm}
\usage{
pelt(
  data,
  type_changement = "mean",
  penalty = "Manual",
  pen.value = var(data) * log(length(data))
)
}
\arguments{
\item{data}{Numeric vector containing the data to segment.}

\item{type_changement}{Type of change to detect. Maybe "mean","var" or "meanvar".}

\item{penalty}{Method of penalty. Maybe "Manual" or "BIC" for example.}

\item{pen.value}{if penalty is "Manual", this is the value of the penalty.}
}
\value{
A numeric vector containing the segmentation points.
}
\description{
Change Point Detection using pelt algorithm
}
\examples{
data <- c(rnorm(40,0,2),rnorm(60,5,2),rnorm(60,1,2),rnorm(40,5,2))
type_changement = "mean"
penalty = "Manual"
pen.value = var(data) * log(length(data))
pelt(data, type_changement, penalty, pen.value )
}
